Revamp GApps scripts
1. Add BRANCH variable for git reset
2. Add log support

+++ GApps/gapps.sh
@@ -31,12 +31,17 @@ if [ "${TYPE}" == "banks" ]
 then
     SOURCEDIR=${ANDROIDDIR}/GApps/Banks
     ZIPBEG=BaNkS
+    BRANCH=m
 elif [ "${TYPE}" == "pn" ]
 then
     SOURCEDIR=${ANDROIDDIR}/GApps/PN
     ZIPBEG=PureNexus
+    BRANCH=mm2
 fi
-ZIPMOVE=~/shared/GApps
+ZIPMOVE=${HOME}/shared/GApps
+# Export the COMPILE_LOG variable for other files to use (I currently handle this via .bashrc)
+# export LOGDIR=${ANDROID_DIR}/Logs
+# export COMPILE_LOG=${LOGDIR}/compile_log_`date +%m_%d_%y`.log
 
 
 
@@ -45,14 +50,23 @@ clear
 
 
 
+# Start tracking time
+START=$(date +%s)
+
+
+
 # Go into repo folder
 cd ${SOURCEDIR}
 
 
 
-# Clean unsaved changes and get new changes
-git reset --hard
-git clean -f -d
+# Clean up repo
+git reset --hard origin/${BRANCH}
+git clean -f -d -x
+
+
+
+# Get new changes
 git pull
 
 
@@ -62,9 +76,23 @@ git pull
 
 
 
-# Remove current GApps and move the new ones in their place
-rm ${ZIPMOVE}/${ZIPBEG}*.zip
-mv ${SOURCEDIR}/out/${ZIPBEG}*.zip ${ZIPMOVE}
+# If the above was successful
+if [ `ls ${SOURCEDIR}/out/${ZIPBEG}*.zip 2>/dev/null | wc -l` != "0" ]
+then
+   BUILD_RESULT_STRING="BUILD SUCCESSFUL"
+
+
+   # Remove current GApps and move the new ones in their place
+   rm ${ZIPMOVE}/${ZIPBEG}*.zip
+   mv ${SOURCEDIR}/out/${ZIPBEG}*.zip ${ZIPMOVE}
+
+
+
+# If the build failed, add a variable
+else
+   BUILD_RESULT_STRING="BUILD FAILED"
+
+fi
 
 
 
@@ -73,13 +101,27 @@ mv ${SOURCEDIR}/out/${ZIPBEG}*.zip ${ZIPMOVE}
 
 
 
+# Stop tracking time
+END=$(date +%s)
+
+
+
 # Go home and we're done!
 cd ${HOME}
 
+
+
 echo -e ${BLDGREEN}
-echo -e "---------------------------------"
-echo -e "COMPILATION AND UPLOAD SUCCESSFUL"
-echo -e "---------------------------------"
+echo -e "-------------------------------------"
+echo -e "SCRIPT ENDING AT $(date +%D\ %r)"
+echo -e ""
+echo -e "${BUILD_RESULT_STRING}!"
+echo -e "TIME: $(echo $((${END}-${START})) | awk '{print int($1/60)" MINUTES AND "int($1%60)" SECONDS"}')"
+echo -e "-------------------------------------"
 echo -e ${RST}
+
+# Add line to compile log
+echo -e "`date +%H:%M:%S`: ${BASH_SOURCE} ${TYPE}" >> ${COMPILE_LOG}
+echo -e "${BUILD_RESULT_STRING} IN $(echo $((${END}-${START})) | awk '{print int($1/60)" MINUTES AND "int($1%60)" SECONDS"}')\n" >> ${COMPILE_LOG}
 
 echo -e "\a"
